Add basis for entity/component system
Add a basis for an entity/component system. Still needs a lot of
fixing on many parts, but already works nicely with template-based
operations on adding, removing and obtaining references to components,
given an entity. Although it uses OO in some parts, specially for
defining new components or an entity, it is still kept to a minimum.

An `Entity` control is nothing more than a wrapper, which can also be
passed by value anywhere, which wraps an entity's global ID and
performs operations on it. It cannot be overriden either, to make sure
that the end user does not try to declare a specific class for a game
object. The entities, though, still need a special control for their
management, which should interact with an underlying data structure.

The documentation rules are updated as well, so that the documentation
lies on header files, but still keep themselves nice and clean to the
eyes, should it need to be read.

C++ version is also bumped to C++17 so it can make use of the latest
available features, such as `std::optional`.