Add work-in-progress scripting component and cmake-ide support
There is a new `.dir-locals.el` file which sets up some variables for
`cmake-ide` on Emacs. It shouldn't be a cause of worrying, since it
tries to deduce the relative path to the project and build
directories, however it includes `eval` forms, which will cause a
warning in Emacs when opened for the first time (and on subsequent
opens, if not marked as safe). Though it would be better to use the
alist on this file itself to set these variables, this setup works
best if one doesn't want to worry about this file's portability.

A preliminar component for a scripted Entity has been added. It still
has a lot of features missing. Depends on Chicken Scheme's API, which
was chosen as extension language for this engine, however the project
is being managed in a way that it does not depend on Chicken Scheme
support on the target system, dropping the requirement and the
component code altogether if Chicken isn't found. Plus, the very
useful CMake scripts to find Chicken on system were copied from a
reliable Chicken Scheme mirror, which is show in each files's
beginning.

Build instructions were also added to README.md for Emacs using
`cmake-ide`.